Create README.md
The Scientific Calculator In Python is a simple project developed using Python. The project file contains a python script (my_Calculator.py). This is a simple GUI-base project which is very easy to understand and use. Also, this project makes a convenient way for the user to gain an idea of how to perform numerical calculations.

The project contains the numbers, operators, and signs like on a normal calculator. So, the user can enter any number or can just click on the numbers they want in the calculations. Hence, the user can use a simple calculator from this application.